Change flag schema for rollout to rollout_percentage, and its type to integer between 1 and 1000.
