+++ server/app.js
@@ -40,5 +40,4 @@ const cleanup = async () => {
   await require('./lib/nats');
 })();
 
-process.on('SIGINT', cleanup);
 process.on('SIGTERM', cleanup);
