+++ server/http/flags.http
@@ -1,12 +1,12 @@
-@appId = 9
-@flagId = 16
+@appId = 1
+@flagId = 1
 ###
 # Create flag
 POST http://localhost:3000/api/apps/{{appId}}/flags
 Content-Type: application/json
 
 {
-  "title": "Flag in app 9 number 1"
+  "title": "Flag in app 1 number 1"
 }
 
 
@@ -31,7 +31,7 @@ PATCH http://localhost:3000/api/flags/{{flagId}}
 Content-Type: application/json
 
 {
-  "error_threshold": 5
+  "rollout_percentage": 550
 }
